- Refactor for db.cpp, db.h. - Make code for map loading and saving more clear. - Add PNG, JPG, DDS, BMP support for artedit using KPLIB. - Add a new class for processing parallax sectors (xmpsky). - Add GUI based import wizard tool. - Add config version check. - Get rid off qloadboard, qsaveboard functions. - Make movement keys for qavedit work a bit slower. - Increase max files from 512 to 1024 per directory for filebrowser. - Add char enumerator (enumstrGetChar). - Some GUI updates. - Misc changes.

+++ src_blood/xmpview.cpp
@@ -47,7 +47,7 @@ inline int QRandom2(int a1)
     return mulscale14(qrand(), a1)-a1;
 }
 
-void viewDoQuake(int strength, int* x, int* y, int* z, int* a, int* h) {
+void viewDoQuake(int strength, int* x, int* y, int* z, short* a, int* h) {
 	
 	if (!strength)
 		return;
